pyproject.toml
[tool.poetry]
name = "reddit-collector"
version = "0.1.0"
description = "Data collector from Reddit."
authors = ["jordaoa <andrejordao50@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
praw = "7.7.0"
python-dotenv = "^1.0.0"
pendulum = "^2.1.2"
apache-airflow = "^2.6.3"
nltk = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
